pyproject.toml
[tool.poetry]
name = "web_server"
version = "0.1.0"
description = "Simple web server stub for testing purposes"
authors = ["Anton Kachurin <anton.kachurin@t-systems.com>"]
license = "Apache-2.0"

[[tool.poetry.source]]
name = "outcatcher"
url = "https://pypi.outcatcher.com/simple/"

[tool.poetry.dependencies]
python = "^3.7"

peewee = "^3.9"
flask = "^1.1"
flask-restful = "^0.3.7"
wsgiserver = "^1.3"
psycopg2 = "^2.8"
python-daemon = "^2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
